+++ include/libqbuttongroup.h
@@ -35,13 +35,13 @@ void* q_buttongroup_metacast(void* self, const char* param1);
 /// @param param1 enum QMetaObject__Call
 /// @param param2 int
 /// @param param3 void*
-int32_t q_buttongroup_metacall(void* self, int64_t param1, int param2, void* param3);
+int32_t q_buttongroup_metacall(void* self, int32_t param1, int param2, void* param3);
 
 /// Allows for overriding the related default method
 ///
 /// @param self QButtonGroup*
 /// @param callback int32_t fn(QButtonGroup*, enum QMetaObject__Call, int, void*)
-void q_buttongroup_on_metacall(void* self, int32_t (*callback)(void*, int64_t, int, void*));
+void q_buttongroup_on_metacall(void* self, int32_t (*callback)(void*, int32_t, int, void*));
 
 /// Base class method implementation
 ///
@@ -49,7 +49,7 @@ void q_buttongroup_on_metacall(void* self, int32_t (*callback)(void*, int64_t, i
 /// @param param1 enum QMetaObject__Call
 /// @param param2 int
 /// @param param3 void*
-int32_t q_buttongroup_qbase_metacall(void* self, int64_t param1, int param2, void* param3);
+int32_t q_buttongroup_qbase_metacall(void* self, int32_t param1, int param2, void* param3);
 
 /// [Qt documentation](https://doc.qt.io/qt-6/qobject.html#tr)
 ///
@@ -327,7 +327,7 @@ void q_buttongroup_kill_timer(void* self, int id);
 ///
 /// @param self QButtonGroup*
 /// @param id enum Qt__TimerId
-void q_buttongroup_kill_timer2(void* self, int64_t id);
+void q_buttongroup_kill_timer2(void* self, int32_t id);
 
 /// Inherited from QObject
 ///
@@ -504,7 +504,7 @@ bool q_buttongroup_move_to_thread2(void* self, void* thread, void* param2);
 /// @param self QButtonGroup*
 /// @param interval int
 /// @param timerType enum Qt__TimerType
-int32_t q_buttongroup_start_timer22(void* self, int interval, int64_t timerType);
+int32_t q_buttongroup_start_timer22(void* self, int interval, int32_t timerType);
 
 /// Inherited from QObject
 ///
@@ -515,7 +515,7 @@ int32_t q_buttongroup_start_timer22(void* self, int interval, int64_t timerType)
 /// @param receiver QObject*
 /// @param method QMetaMethod*
 /// @param typeVal enum Qt__ConnectionType
-QMetaObject__Connection* q_buttongroup_connect5(void* sender, void* signal, void* receiver, void* method, int64_t typeVal);
+QMetaObject__Connection* q_buttongroup_connect5(void* sender, void* signal, void* receiver, void* method, int32_t typeVal);
 
 /// Inherited from QObject
 ///
@@ -526,7 +526,7 @@ QMetaObject__Connection* q_buttongroup_connect5(void* sender, void* signal, void
 /// @param signal const char*
 /// @param member const char*
 /// @param typeVal enum Qt__ConnectionType
-QMetaObject__Connection* q_buttongroup_connect4(void* self, void* sender, const char* signal, const char* member, int64_t typeVal);
+QMetaObject__Connection* q_buttongroup_connect4(void* self, void* sender, const char* signal, const char* member, int32_t typeVal);
 
 /// Inherited from QObject
 ///
